Telemetry performance and bug fix extravaganza
Telemetry now covers more things, notably the (SDL) event loop.
Added support for unsigned 64bit integers to NetworkBuffer.
Better buffer usage in AssetProtocolHandler.
AssetProtocolHandlers can now be freed properly.
loom_newArray allocation now skips constructors for fundamental types.
Telemetry now uses an event-and-lookup design, which is a bit tougher on bandwidth, but much lower on processing. The bandwidth issue is side-stepped with new buffered network writes.
Network writes are now buffered in a bip buffer. This avoids stalls and blocking.
The asset agent connection is now more aware of issues on both ends, reads and writes now time out after some time and the connection is dropped.
Changed from double to long long in readTimerNano to avoid unnecessary FP ops.
Fixed some JSON leaks and simplified implementation / memory management wrt usage a bit.
Added a `useCache` parameter to `utByteArray` to expose the underlying `utArray` cache.
Added `utByteArray` reading/writing unsigned 64-bit ints.
Added temp fix for reading/writing unaligned `utByteArray` strings.
Added uthash, a simple hash table for use from C code.
Added `utString::alloc` to provide a way to reserve a string buffer in advance.
Added an unsigned 64-bit int hash key support to `utHashTable`.
Changed some SQLite code to use the newer JSON APIs that manage memory automatically.
Fixed texture mipmapping code to actually allocate properly.
Fixed some internal string functions to use the lua string length and not recompute all the time with strlen, noticable with larger strings.
Fixed a bool warning for `openURL`.
Made a `MemberInfo` check function faster by avoiding string length computation.
Added `LoomProfilerRoot` to `MethodBase` to support the new telemetry system.
Fixed some telemetry server memory leaks.
Added a ping timeout to the asset agent connection.
Added support for multiple "root" tick ranges to the telemetry client.
Added display of write/read overhead to telemetry client.

+++ loom/common/platform/platformTime.h
@@ -34,7 +34,7 @@ typedef void * loom_precision_timer_t;
 loom_precision_timer_t loom_startTimer();
 void loom_resetTimer(loom_precision_timer_t timer);
 int loom_readTimer(loom_precision_timer_t timer);
-double loom_readTimerNano(loom_precision_timer_t timer);
+long long loom_readTimerNano(loom_precision_timer_t timer);
 void loom_destroyTimer(loom_precision_timer_t timer);
 
 #ifdef __cplusplus
